requirements.txt
selenium>=4.15.0
flask>=3.0.0
gunicorn>=21.0.0
webdriver-manager>=4.0.0
python-dotenv>=1.0.0
